# By liurui12w
# Via liurui12w
* 'master' of https://github.com/dreamlx/meimeishop:
  fix: fix add_record api fix
  fix: fix add_record api fix
  fix
  fix: 赵翔需求api 时间精度fix
  fix: 供应详情api fix,增加评价model
  fix: t_list api fix 2
  fix: t-list api
  refactor: 供应方api补充分类字段
  feat:测试功能 提供给兆祥使用,一个get一个post api
  refactor: 供需api修改,可查看某人的供需数据
  refactor: 供应方需求方,查看全部数据,只能修改删除自己创建的数据
  fix: 用户绑定根据手机号创建相对应账户

+++ app/controllers/api/small_routine/providers_controller.rb
@@ -5,7 +5,14 @@ class Api::SmallRoutine::ProvidersController < Api::SmallRoutine::BaseController
   def index
     page = params[:page] || 1
     per = params[:per] || 20
-    record = Provider.where(user_id: @current_wx_user.id).order("created_at desc")
+    type = params[:type] 
+    if type == "all"
+      record = Provider.all.order("created_at desc")
+    elsif type == "own"
+      record = Provider.where(user_id: @current_wx_user.user_id).order("created_at desc")
+    else
+      record = Provider.where(user_id: params[:user_id]).order("created_at desc")
+    end
     @record = Kaminari.paginate_array(record).page(page).per(per)
   end
 
@@ -28,6 +35,10 @@ class Api::SmallRoutine::ProvidersController < Api::SmallRoutine::BaseController
   end
 
   def update
+    if @current_wx_user.user_id != @record.user_id
+      return render json: {status: 400, message: "不是本人创建,无法修改"}
+    end
+    
     if @record.update!(provider_params)
       result = [200, '修改成功']
     else
@@ -37,6 +48,10 @@ class Api::SmallRoutine::ProvidersController < Api::SmallRoutine::BaseController
   end
 
   def destroy
+    if @current_wx_user.user_id != @record.user_id
+      return render json: {status: 400, message: "不是本人创建,无法删除"}
+    end
+    
     if @record.destroy
       result = [200, '删除成功']
     else
